+++ Kaa/RecordCountWithTimeLimitLogUploadStrategy.h
@@ -10,10 +10,11 @@
 
 #import "DefaultLogUploadStrategy.h"
 #import "TimeCommons.h"
+
 @interface RecordCountWithTimeLimitLogUploadStrategy : DefaultLogUploadStrategy
 
-@property (nonatomic) long lastUploadTime;
+@property (nonatomic) int64_t lastUploadTime;
 
-- (instancetype) initWithCountThreshold:(NSInteger)countThreshold TimeLimit:(long)timeLimit andTimeUnit:(TimeUnit)timeUnit;
+- (instancetype)initWithCountThreshold:(int32_t)countThreshold TimeLimit:(int64_t)timeLimit andTimeUnit:(TimeUnit)timeUnit;
 
 @end
